Refactor to physics constraints API
This refactor is intended to allow multiple constraints in the same
rigid body.

Main points in this refactor:

- Constraints are no longer GVRComponent;
- constraints constructor receives both rigid bodies;
- constraint construction is no longer finished only when added to world;
- GVRWorld now keeps reference to each added constraint;
- physics loader uses new 'mark' flag to identify when a constraint must
  not be deleted.

+++ GVRf/Extensions/gvrf-physics/src/main/jni/engine/physics_loader_jni.cpp
@@ -66,11 +66,11 @@ Java_org_gearvrf_physics_NativePhysics3DLoader_ctor(JNIEnv* env, jclass clazz,
     }
 
     char *buffer = new char[len];
-    memcpy(buffer, data, len);
+    memcpy(buffer, data, (size_t)len);
 
     env->ReleaseByteArrayElements(byteArr, data, JNI_ABORT);
 
-    PhysicsLoader *loader = new BulletFileLoader(buffer, len, ignoreUpAxis);
+    PhysicsLoader *loader = new BulletFileLoader(buffer, (size_t)len, ignoreUpAxis);
 
     delete[] buffer;
